+++ src/lib.rs
@@ -20,7 +20,7 @@ By default, a TUI is provided to visualize all state. Have a look at [the exampl
 
 # Logging
 
-If the feature `log-renderer` is enabled (default), most calls to `progress` will also be logged.
+If the feature `progress-tree-log` is enabled (default), most calls to `progress` will also be logged.
 That way, even without a terminal user interface, there will be progress messages.
 Please note that logging to stdout should not be performed with this feature enabled and a terminal user interface is shown, as this will
 seriously interfere with the TUI.
@@ -33,14 +33,17 @@ Please have a look at the [dashboard demo](https://github.com/Byron/crates-io-cl
 
 Run it with `cargo run --example dashboard` and see what else it can do by checking out `cargo run --example dashboard -- --help`.
 */
+#[cfg(feature = "progress-tree")]
 pub mod tree;
+#[cfg(feature = "progress-tree")]
+#[doc(inline)]
 pub use tree::{Options as TreeOptions, Root as Tree};
 
 pub mod render;
 
-#[cfg(feature = "log-renderer")]
+#[cfg(feature = "progress-tree-log")]
 pub use log::info;
-#[cfg(feature = "log-renderer")]
+#[cfg(feature = "progress-tree-log")]
 pub use log::warn;
 
 #[cfg(any(feature = "humantime", feature = "time"))]
@@ -59,7 +62,7 @@ pub use traits::{Progress, Root};
 mod throughput;
 pub use crate::throughput::Throughput;
 
-#[cfg(not(feature = "log-renderer"))]
+#[cfg(not(feature = "progress-tree-log"))]
 mod log {
     /// Stub
     #[macro_export(local_inner_macros)]